feat: PeriodicTableコンポーネントと関連データ、スタイルを削除し、インポートパスを修正
