--- src/components/periodic_table/data/elementData.ts
@@ -1,722 +1,0 @@
-export const periodicTableData = [
-  // Period 1
-  {
-    symbol: "H",
-    name: "水素",
-    atomicNumber: 1,
-    category: "nonmetal",
-    color: "lightblue",
-  },
-
-  // placeholder
-  ...Array(16).fill({
-    symbol: "",
-    name: "",
-    atomicNumber: "",
-    category: "placeholder",
-    color: "",
-  }),
-
-  {
-    symbol: "He",
-    name: "ヘリウム",
-    atomicNumber: 2,
-    category: "noble-gas",
-    color: "plum",
-  },
-
-  // Period 2
-  {
-    symbol: "Li",
-    name: "リチウム",
-    atomicNumber: 3,
-    category: "alkali-metal",
-    color: "lightcoral",
-  },
-  {
-    symbol: "Be",
-    name: "ベリリウム",
-    atomicNumber: 4,
-    category: "alkaline-earth",
-    color: "lightsalmon",
-  },
-
-  // placeholder
-  ...Array(10).fill({
-    symbol: "",
-    name: "",
-    atomicNumber: "",
-    category: "placeholder",
-    color: "",
-  }),
-  {
-    symbol: "B",
-    name: "ホウ素",
-    atomicNumber: 5,
-    category: "metalloid",
-    color: "lightseagreen",
-  },
-  {
-    symbol: "C",
-    name: "炭素",
-    atomicNumber: 6,
-    category: "nonmetal",
-    color: "lightblue",
-  },
-  {
-    symbol: "N",
-    name: "窒素",
-    atomicNumber: 7,
-    category: "nonmetal",
-    color: "lightblue",
-  },
-  {
-    symbol: "O",
-    name: "酸素",
-    atomicNumber: 8,
-    category: "nonmetal",
-    color: "lightblue",
-  },
-  {
-    symbol: "F",
-    name: "フッ素",
-    atomicNumber: 9,
-    category: "halogen",
-    color: "lightsteelblue",
-  },
-  {
-    symbol: "Ne",
-    name: "ネオン",
-    atomicNumber: 10,
-    category: "noble-gas",
-    color: "plum",
-  },
-
-  // Period 3
-  {
-    symbol: "Na",
-    name: "ナトリウム",
-    atomicNumber: 11,
-    category: "alkali-metal",
-    color: "lightcoral",
-  },
-  {
-    symbol: "Mg",
-    name: "マグネシウム",
-    atomicNumber: 12,
-    category: "alkaline-earth",
-    color: "lightsalmon",
-  },
-
-  // placeholder
-  ...Array(10).fill({
-    symbol: "",
-    name: "",
-    atomicNumber: "",
-    category: "placeholder",
-    color: "",
-  }),
-
-  {
-    symbol: "Al",
-    name: "アルミニウム",
-    atomicNumber: 13,
-    category: "post-transition-metal",
-    color: "lightgreen",
-  },
-  {
-    symbol: "Si",
-    name: "ケイ素",
-    atomicNumber: 14,
-    category: "metalloid",
-    color: "lightseagreen",
-  },
-  {
-    symbol: "P",
-    name: "リン",
-    atomicNumber: 15,
-    category: "nonmetal",
-    color: "lightblue",
-  },
-  {
-    symbol: "S",
-    name: "硫黄",
-    atomicNumber: 16,
-    category: "nonmetal",
-    color: "lightblue",
-  },
-  {
-    symbol: "Cl",
-    name: "塩素",
-    atomicNumber: 17,
-    category: "halogen",
-    color: "lightsteelblue",
-  },
-  {
-    symbol: "Ar",
-    name: "アルゴン",
-    atomicNumber: 18,
-    category: "noble-gas",
-    color: "plum",
-  },
-
-  // Period 4
-  {
-    symbol: "K",
-    name: "カリウム",
-    atomicNumber: 19,
-    category: "alkali-metal",
-    color: "lightcoral",
-  },
-  {
-    symbol: "Ca",
-    name: "カルシウム",
-    atomicNumber: 20,
-    category: "alkaline-earth",
-    color: "lightsalmon",
-  },
-  {
-    symbol: "Sc",
-    name: "スカンジウム",
-    atomicNumber: 21,
-    category: "transition-metal",
-    color: "lightgoldenrodyellow",
-  },
-  {
-    symbol: "Ti",
-    name: "チタン",
-    atomicNumber: 22,
-    category: "transition-metal",
-    color: "lightgoldenrodyellow",
-  },
-  {
-    symbol: "V",
-    name: "バナジウム",
-    atomicNumber: 23,
-    category: "transition-metal",
-    color: "lightgoldenrodyellow",
-  },
-  {
-    symbol: "Cr",
-    name: "クロム",
-    atomicNumber: 24,
-    category: "transition-metal",
-    color: "lightgoldenrodyellow",
-  },
-  {
-    symbol: "Mn",
-    name: "マンガン",
-    atomicNumber: 25,
-    category: "transition-metal",
-    color: "lightgoldenrodyellow",
-  },
-  {
-    symbol: "Fe",
-    name: "鉄",
-    atomicNumber: 26,
-    category: "transition-metal",
-    color: "lightgoldenrodyellow",
-  },
-  {
-    symbol: "Co",
-    name: "コバルト",
-    atomicNumber: 27,
-    category: "transition-metal",
-    color: "lightgoldenrodyellow",
-  },
-  {
-    symbol: "Ni",
-    name: "ニッケル",
-    atomicNumber: 28,
-    category: "transition-metal",
-    color: "lightgoldenrodyellow",
-  },
-  {
-    symbol: "Cu",
-    name: "銅",
-    atomicNumber: 29,
-    category: "transition-metal",
-    color: "lightgoldenrodyellow",
-  },
-  {
-    symbol: "Zn",
-    name: "亜鉛",
-    atomicNumber: 30,
-    category: "transition-metal",
-    color: "lightgoldenrodyellow",
-  },
-  {
-    symbol: "Ga",
-    name: "ガリウム",
-    atomicNumber: 31,
-    category: "post-transition-metal",
-    color: "lightgreen",
-  },
-  {
-    symbol: "Ge",
-    name: "ゲルマニウム",
-    atomicNumber: 32,
-    category: "metalloid",
-    color: "lightseagreen",
-  },
-  {
-    symbol: "As",
-    name: "ヒ素",
-    atomicNumber: 33,
-    category: "metalloid",
-    color: "lightseagreen",
-  },
-  {
-    symbol: "Se",
-    name: "セレン",
-    atomicNumber: 34,
-    category: "nonmetal",
-    color: "lightblue",
-  },
-  {
-    symbol: "Br",
-    name: "臭素",
-    atomicNumber: 35,
-    category: "halogen",
-    color: "lightsteelblue",
-  },
-  {
-    symbol: "Kr",
-    name: "クリプトン",
-    atomicNumber: 36,
-    category: "noble-gas",
-    color: "plum",
-  },
-
-  // Period 5
-  {
-    symbol: "Rb",
-    name: "ルビジウム",
-    atomicNumber: 37,
-    category: "alkali-metal",
-    color: "lightcoral",
-  },
-  {
-    symbol: "Sr",
-    name: "ストロンチウム",
-    atomicNumber: 38,
-    category: "alkaline-earth",
-    color: "lightsalmon",
-  },
-  {
-    symbol: "Y",
-    name: "イットリウム",
-    atomicNumber: 39,
-    category: "transition-metal",
-    color: "lightgoldenrodyellow",
-  },
-  {
-    symbol: "Zr",
-    name: "ジルコニウム",
-    atomicNumber: 40,
-    category: "transition-metal",
-    color: "lightgoldenrodyellow",
-  },
-  {
-    symbol: "Nb",
-    name: "ニオブ",
-    atomicNumber: 41,
-    category: "transition-metal",
-    color: "lightgoldenrodyellow",
-  },
-  {
-    symbol: "Mo",
-    name: "モリブデン",
-    atomicNumber: 42,
-    category: "transition-metal",
-    color: "lightgoldenrodyellow",
-  },
-  {
-    symbol: "Tc",
-    name: "テクネチウム",
-    atomicNumber: 43,
-    category: "transition-metal",
-    color: "lightgoldenrodyellow",
-  },
-  {
-    symbol: "Ru",
-    name: "ルテニウム",
-    atomicNumber: 44,
-    category: "transition-metal",
-    color: "lightgoldenrodyellow",
-  },
-  {
-    symbol: "Rh",
-    name: "ロジウム",
-    atomicNumber: 45,
-    category: "transition-metal",
-    color: "lightgoldenrodyellow",
-  },
-  {
-    symbol: "Pd",
-    name: "パラジウム",
-    atomicNumber: 46,
-    category: "transition-metal",
-    color: "lightgoldenrodyellow",
-  },
-  {
-    symbol: "Ag",
-    name: "銀",
-    atomicNumber: 47,
-    category: "transition-metal",
-    color: "lightgoldenrodyellow",
-  },
-  {
-    symbol: "Cd",
-    name: "カドミウム",
-    atomicNumber: 48,
-    category: "transition-metal",
-    color: "lightgoldenrodyellow",
-  },
-  {
-    symbol: "In",
-    name: "インジウム",
-    atomicNumber: 49,
-    category: "post-transition-metal",
-    color: "lightgreen",
-  },
-  {
-    symbol: "Sn",
-    name: "スズ",
-    atomicNumber: 50,
-    category: "post-transition-metal",
-    color: "lightgreen",
-  },
-  {
-    symbol: "Sb",
-    name: "アンチモン",
-    atomicNumber: 51,
-    category: "metalloid",
-    color: "lightseagreen",
-  },
-  {
-    symbol: "Te",
-    name: "テルル",
-    atomicNumber: 52,
-    category: "metalloid",
-    color: "lightseagreen",
-  },
-  {
-    symbol: "I",
-    name: "ヨウ素",
-    atomicNumber: 53,
-    category: "halogen",
-    color: "lightsteelblue",
-  },
-  {
-    symbol: "Xe",
-    name: "キセノン",
-    atomicNumber: 54,
-    category: "noble-gas",
-    color: "plum",
-  },
-
-  // Period 6
-  {
-    symbol: "Cs",
-    name: "セシウム",
-    atomicNumber: 55,
-    category: "alkali-metal",
-    color: "lightcoral",
-  },
-  {
-    symbol: "Ba",
-    name: "バリウム",
-    atomicNumber: 56,
-    category: "alkaline-earth",
-    color: "lightsalmon",
-  },
-
-  // ランタノイドのスペース
-  {
-    symbol: "",
-    name: "",
-    atomicNumber: "",
-    category: "placeholder",
-    color: "",
-  },
-
-  {
-    symbol: "Hf",
-    name: "ハフニウム",
-    atomicNumber: 72,
-    category: "transition-metal",
-    color: "lightgoldenrodyellow",
-  },
-  {
-    symbol: "Ta",
-    name: "タンタル",
-    atomicNumber: 73,
-    category: "transition-metal",
-    color: "lightgoldenrodyellow",
-  },
-  {
-    symbol: "W",
-    name: "タングステン",
-    atomicNumber: 74,
-    category: "transition-metal",
-    color: "lightgoldenrodyellow",
-  },
-  {
-    symbol: "Re",
-    name: "レニウム",
-    atomicNumber: 75,
-    category: "transition-metal",
-    color: "lightgoldenrodyellow",
-  },
-  {
-    symbol: "Os",
-    name: "オスミウム",
-    atomicNumber: 76,
-    category: "transition-metal",
-    color: "lightgoldenrodyellow",
-  },
-  {
-    symbol: "Ir",
-    name: "イリジウム",
-    atomicNumber: 77,
-    category: "transition-metal",
-    color: "lightgoldenrodyellow",
-  },
-  {
-    symbol: "Pt",
-    name: "白金",
-    atomicNumber: 78,
-    category: "transition-metal",
-    color: "lightgoldenrodyellow",
-  },
-  {
-    symbol: "Au",
-    name: "金",
-    atomicNumber: 79,
-    category: "transition-metal",
-    color: "lightgoldenrodyellow",
-  },
-  {
-    symbol: "Hg",
-    name: "水銀",
-    atomicNumber: 80,
-    category: "transition-metal",
-    color: "lightgoldenrodyellow",
-  },
-  {
-    symbol: "Tl",
-    name: "タリウム",
-    atomicNumber: 81,
-    category: "post-transition-metal",
-    color: "lightgreen",
-  },
-  {
-    symbol: "Pb",
-    name: "鉛",
-    atomicNumber: 82,
-    category: "post-transition-metal",
-    color: "lightgreen",
-  },
-  {
-    symbol: "Bi",
-    name: "ビスマス",
-    atomicNumber: 83,
-    category: "post-transition-metal",
-    color: "lightgreen",
-  },
-  {
-    symbol: "Po",
-    name: "ポロニウム",
-    atomicNumber: 84,
-    category: "metalloid",
-    color: "lightseagreen",
-  },
-  {
-    symbol: "At",
-    name: "アスタチン",
-    atomicNumber: 85,
-    category: "halogen",
-    color: "lightsteelblue",
-  },
-  {
-    symbol: "Rn",
-    name: "ラドン",
-    atomicNumber: 86,
-    category: "noble-gas",
-    color: "plum",
-  },
-
-  // Period 7
-  {
-    symbol: "Fr",
-    name: "フランシウム",
-    atomicNumber: 87,
-    category: "alkali-metal",
-    color: "lightcoral",
-  },
-  {
-    symbol: "Ra",
-    name: "ラジウム",
-    atomicNumber: 88,
-    category: "alkaline-earth",
-    color: "lightsalmon",
-  },
-
-  // アクチノイドのスペース
-  { symbol: "", name: "", atomicNumber: "", category: "placeholder" },
-
-  {
-    symbol: "Rf",
-    name: "ラザホージウム",
-    atomicNumber: 104,
-    category: "transition-metal",
-    color: "lightgoldenrodyellow",
-  },
-  {
-    symbol: "Db",
-    name: "ドブニウム",
-    atomicNumber: 105,
-    category: "transition-metal",
-    color: "lightgoldenrodyellow",
-  },
-  {
-    symbol: "Sg",
-    name: "シーボーギウム",
-    atomicNumber: 106,
-    category: "transition-metal",
-    color: "lightgoldenrodyellow",
-  },
-  {
-    symbol: "Bh",
-    name: "ボーリウム",
-    atomicNumber: 107,
-    category: "transition-metal",
-    color: "lightgoldenrodyellow",
-  },
-  {
-    symbol: "Hs",
-    name: "ハッシウム",
-    atomicNumber: 108,
-    category: "transition-metal",
-    color: "lightgoldenrodyellow",
-  },
-  {
-    symbol: "Mt",
-    name: "マイトネリウム",
-    atomicNumber: 109,
-    category: "unknown",
-    color: "lightskyblue",
-  },
-  {
-    symbol: "Ds",
-    name: "ダームスタチウム",
-    atomicNumber: 110,
-    category: "unknown",
-    color: "lightskyblue",
-  },
-  {
-    symbol: "Rg",
-    name: "レントゲニウム",
-    atomicNumber: 111,
-    category: "unknown",
-    color: "lightskyblue",
-  },
-  {
-    symbol: "Cn",
-    name: "コペルニシウム",
-    atomicNumber: 112,
-    category: "transition-metal",
-    color: "lightgoldenrodyellow",
-  },
-  {
-    symbol: "Nh",
-    name: "ニホニウム",
-    atomicNumber: 113,
-    category: "unknown",
-    color: "lightskyblue",
-  },
-  {
-    symbol: "Fl",
-    name: "フレロビウム",
-    atomicNumber: 114,
-    category: "unknown",
-    color: "lightskyblue",
-  },
-  {
-    symbol: "Mc",
-    name: "モスコビウム",
-    atomicNumber: 115,
-    category: "unknown",
-    color: "lightskyblue",
-  },
-  {
-    symbol: "Lv",
-    name: "リバモリウム",
-    atomicNumber: 116,
-    category: "unknown",
-    color: "lightskyblue",
-  },
-  {
-    symbol: "Ts",
-    name: "テネシン",
-    atomicNumber: 117,
-    category: "halogen",
-    color: "lightsteelblue",
-  },
-  {
-    symbol: "Og",
-    name: "オガネソン",
-    atomicNumber: 118,
-    category: "noble-gas",
-    color: "plum",
-  },
-
-  // Lanthanides
-  ...[
-    { symbol: "La", name: "ランタン" },
-    { symbol: "Ce", name: "セリウム" },
-    { symbol: "Pr", name: "プラセオジム" },
-    { symbol: "Nd", name: "ネオジム" },
-    { symbol: "Pm", name: "プロメチウム" },
-    { symbol: "Sm", name: "サマリウム" },
-    { symbol: "Eu", name: "ユウロピウム" },
-    { symbol: "Gd", name: "ガドリニウム" },
-    { symbol: "Tb", name: "テルビウム" },
-    { symbol: "Dy", name: "ジスプロシウム" },
-    { symbol: "Ho", name: "ホルミウム" },
-    { symbol: "Er", name: "エルビウム" },
-    { symbol: "Tm", name: "ツリウム" },
-    { symbol: "Yb", name: "イッテルビウム" },
-    { symbol: "Lu", name: "ルテチウム" },
-  ].map((el, i) => ({
-    ...el,
-    atomicNumber: 57 + i,
-    category: "lanthanide",
-    color: "lightpink",
-  })),
-
-  // Actinides
-  ...[
-    { symbol: "Ac", name: "アクチニウム" },
-    { symbol: "Th", name: "トリウム" },
-    { symbol: "Pa", name: "プロトアクチニウム" },
-    { symbol: "U", name: "ウラン" },
-    { symbol: "Np", name: "ネプツニウム" },
-    { symbol: "Pu", name: "プルトニウム" },
-    { symbol: "Am", name: "アメリシウム" },
-    { symbol: "Cm", name: "キュリウム" },
-    { symbol: "Bk", name: "バークリウム" },
-    { symbol: "Cf", name: "カリホルニウム" },
-    { symbol: "Es", name: "アインスタイニウム" },
-    { symbol: "Fm", name: "フェルミウム" },
-    { symbol: "Md", name: "メンデレビウム" },
-    { symbol: "No", name: "ノーベリウム" },
-    { symbol: "Lr", name: "ローレンシウム" },
-  ].map((el, i) => ({
-    ...el,
-    atomicNumber: 89 + i,
-    category: "actinide",
-    color: "lightgray",
-  })),
-];
