--- src/components/periodic_table/styles/preriodic_table.css
@@ -1,20 +1,0 @@
-.periodic-table-container {
-  width: 100%;
-  margin: auto;
-  background-color: #222;
-  border-radius: 10px;
-  padding: 40px;
-}
-
-.grid {
-  display: grid;
-  grid-template-columns: repeat(18, 1fr);
-  gap: 5px;
-}
-
-.lanthanide-actinide {
-  margin-top: 10px;
-  display: grid;
-  grid-template-columns: repeat(14, 1fr);
-  gap: 5px;
-}
